Added Cardio Tracker
Added a Cardio Tracker to the app. It's very basic, you just log what type of cardio you did, how long you did it and how many calories you burned according to the machine. For now, cardio types are just typed directly when you log cardio, but I will be adding a dropdown and ability to add types of cardio. Also started a basic Soreness Tracker. The idea is being able to log your soreness the day after a workout to be able to determine how effective it was.