Added functionality to CalendarView
I'm quite happy with how this pursuit is turning out. I'm still a bit embarrassed by the code itself, however the performance is still miles better than any other package I've tried. The calendar now highlights any date with data recorded to it, and when a user selected a date with data recorded to it, they are shown the corresponding DateReviewListItem component.